Update and refactor PlayerScouter
