--- runelite-client/src/main/java/net/runelite/client/plugins/playerscouter/AttackerOverlay.java
@@ -1,79 +1,0 @@
-/*
- * Copyright (c) 2019, ganom <https://github.com/Ganom>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.playerscouter;
-
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics2D;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import net.runelite.client.graphics.ModelOutlineRenderer;
-import net.runelite.client.ui.overlay.Overlay;
-import net.runelite.client.ui.overlay.OverlayLayer;
-import net.runelite.client.ui.overlay.OverlayPosition;
-
-@Singleton
-public class AttackerOverlay extends Overlay
-{
-	private static final Color TRANSPARENT = new Color(0, 0, 0, 0);
-
-	private final PlayerScouter plugin;
-	private final ModelOutlineRenderer outlineRenderer;
-
-	@Inject
-	public AttackerOverlay(final PlayerScouter plugin, final ModelOutlineRenderer outlineRenderer)
-	{
-		this.plugin = plugin;
-		this.outlineRenderer = outlineRenderer;
-		setPosition(OverlayPosition.DYNAMIC);
-		setLayer(OverlayLayer.ABOVE_SCENE);
-	}
-
-	@Override
-	public Dimension render(Graphics2D graphics)
-	{
-		if (!plugin.isOverlayEnabled() || plugin.getPlayerContainer().isEmpty())
-		{
-			return null;
-		}
-
-		plugin.getPlayerContainer().forEach(player ->
-		{
-			if (!player.isTarget())
-			{
-				return;
-			}
-
-			final AttackStyle attackStyle = player.getAttackStyle();
-
-			if (attackStyle.getPrayer() == null)
-			{
-				return;
-			}
-
-			outlineRenderer.drawOutline(player.getPlayer(), 2, attackStyle.getColor(), TRANSPARENT);
-		});
-		return null;
-	}
-}
