+++ runelite-client/src/main/java/net/runelite/client/plugins/playerscouter/PlayerScouterConfig.java
@@ -42,21 +42,10 @@ public interface PlayerScouterConfig extends Config
 	}
 
 	@ConfigItem(
-		keyName = "overlayEnabled",
-		name = "Attacker Overlay",
-		description = "This will highlight your attacker.",
-		position = 1
-	)
-	default boolean overlayEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
 		keyName = "onlyWildy",
 		name = "Only Scout in Wildy",
 		description = "This will only scout players in the wilderness.",
-		position = 2
+		position = 1
 	)
 	default boolean onlyWildy()
 	{
@@ -67,7 +56,7 @@ public interface PlayerScouterConfig extends Config
 		keyName = "outputItems",
 		name = "Output Items",
 		description = "This will output all of their risked gear to the webhook.",
-		position = 3
+		position = 2
 	)
 	default boolean outputItems()
 	{
@@ -78,7 +67,7 @@ public interface PlayerScouterConfig extends Config
 		keyName = "minimumRisk",
 		name = "Minimum Risk",
 		description = "Minimum risk for the player to be scouted.",
-		position = 4
+		position = 3
 	)
 	default int minimumRisk()
 	{
@@ -89,7 +78,7 @@ public interface PlayerScouterConfig extends Config
 		keyName = "minimumValue",
 		name = "Minimum Value",
 		description = "Minimum value for the item to be posted on discord.",
-		position = 5
+		position = 4
 	)
 	default int minimumValue()
 	{
@@ -100,7 +89,7 @@ public interface PlayerScouterConfig extends Config
 		keyName = "timeout",
 		name = "Timeout",
 		description = "Minimum amount of ticks before the player can be scouted again. (1 tick = 600ms)",
-		position = 6
+		position = 5
 	)
 	default int timeout()
 	{
